acesso restrito a admin, nova home

+++ app/Http/Controllers/ContatosController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Contato;
 use Session;
-use Illuminate\Support\Facades\Auth;
 
 class ContatosController extends Controller
 {
@@ -33,7 +33,7 @@ class ContatosController extends Controller
 
     public function create()
     {
-        if(Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
             return view('contato.create');
         } else {
             return redirect('login');
@@ -48,10 +48,10 @@ class ContatosController extends Controller
      */
     public function store(Request $request)
     {
-        if(Auth::check()){
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
             $this->validate($request,[
-            'nome' => 'required|min:3',
-            'email' => 'required|e-mail',
+                'nome' => 'required|min:3',
+                'email' => 'required|e-mail',
                 'telefone' => 'required',
                 'cidade' => 'required',
                 'estado' => 'required',
@@ -96,7 +96,7 @@ class ContatosController extends Controller
      */
     public function edit($id)
     {
-        if(Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {            
             $contato = Contato::find($id);
             return view('contato.edit', array('contato' => $contato));
         } else {
@@ -113,7 +113,7 @@ class ContatosController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if(Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {            
             $this->validate($request,[
                 'nome'=>'required|min:3',
                 'email'=>'required|email',
@@ -149,7 +149,7 @@ class ContatosController extends Controller
      */
     public function destroy(Request $request, $id)
     {
-        if(Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {            
             $contato = Contato::find($id);
             if (isset($request->foto)) {
                 unlink($request->foto);
